Moved around files to make it so there's no longer a direct dependency needed for ALCProxy.Communication
