+++ tests/ALCProxy.Tests/ALCProxyTests.cs
@@ -8,7 +8,6 @@ using System.IO;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.Loader;
-using ALCProxy.Communication;
 using ALCProxy.Proxy;
 using Xunit;
